Fix cell linking: 'strict' flag must be removed
When adding new cell links, the 'strict' flag must not be present on the cell object. Values of both 'true' and 'false' will cause the API call the fail and the error message does not correctly reflect the issue. This will detect cell links objects and force the 'strict' flag to be removed.